patrial donr
Ok so now the calendar successfully syncs and shows both personal and shared events, but doesn'y currently sync events in the future. only shows events from the past. this os only one way, as if you add an event for the future in the app, it shows in the google calendar. but adding an event in the google calendar and then syncing the app does not get it to show in the app or sync, as they aren't in the firestore DB either